+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## v0.0.2 - in progress
 ### Added
 - Mirror Globus directory to local cache
-
+- Fix `--type_metadata` so it still works without a submission directory.
 
 ## [v0.0.1](https://github.com/hubmapconsortium/ingest-validation-tools/tree/v0.0.1) - 2020-04-13
 ### Added